refactor: error 패키지 및 객체 모두 exception으로 변경
